# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface OpenaiConversation {
  id: string;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}

export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type TicketInputLanguage = typeof TicketInputLanguage[keyof typeof TicketInputLanguage];


export const TicketInputLanguage = {
  en: 'en',
  ar: 'ar',
} as const;

export type TicketInputTranscriptItemRole = typeof TicketInputTranscriptItemRole[keyof typeof TicketInputTranscriptItemRole];


export const TicketInputTranscriptItemRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export type TicketInputTranscriptItem = {
  role: TicketInputTranscriptItemRole;
  /** @maxLength 4000 */
  content: string;
};

export interface TicketInput {
  /**
     * @minLength 3
     * @maxLength 320
     */
  email: string;
  /**
     * @maxLength 200
     * @nullable
     */
  name?: string | null;
  /** @nullable */
  conversationId?: string | null;
  language?: TicketInputLanguage;
  /**
     * @minItems 1
     * @maxItems 60
     */
  transcript: TicketInputTranscriptItem[];
}

export interface Ticket {
  ticketNumber: string;
  subject: string;
  createdAt: string;
}

/**
 * Website language the visitor was using; the confirmation email is sent in this language
 */
export type LeadInputLanguage = typeof LeadInputLanguage[keyof typeof LeadInputLanguage];


export const LeadInputLanguage = {
  en: 'en',
  ar: 'ar',
} as const;

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  whatsapp: string;
  /** @nullable */
  employees?: number | null;
  /** @nullable */
  tier?: string | null;
  /** @nullable */
  tierPrice?: number | null;
  /** @nullable */
  subscription?: number | null;
  /** @nullable */
  totalReturn?: number | null;
  /** Website language the visitor was using; the confirmation email is sent in this language */
  language?: LeadInputLanguage;
}

export interface Lead {
  id: number;
  name: string;
  company: string;
  email: string;
  whatsapp: string;
  /** @nullable */
  employees?: number | null;
  /** @nullable */
  tier?: string | null;
  /** @nullable */
  tierPrice?: number | null;
  /** @nullable */
  subscription?: number | null;
  /** @nullable */
  totalReturn?: number | null;
  createdAt: string;
}