# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Design partner lead capture
  - name: openai
    description: OpenAI AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /lead:
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit a design partner lead
      description: Captures a design partner application with a calculator snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /ticket:
    post:
      operationId: createTicket
      tags: [leads]
      summary: Create a support ticket from a chat conversation
      description: Summarizes the chat conversation, stores a ticket, and emails both the visitor and the Nizamy team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /openai/conversations:
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
components:
  schemas:
    OpenaiConversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: string
          format: uuid
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
      required:
        - content
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TicketInput:
      type: object
      required:
        - email
        - transcript
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 320
        name:
          type: ["string", "null"]
          maxLength: 200
        conversationId:
          type: ["string", "null"]
          format: uuid
        language:
          type: string
          enum: [en, ar]
          default: ar
        transcript:
          type: array
          minItems: 1
          maxItems: 60
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
                maxLength: 4000
    Ticket:
      type: object
      required:
        - ticketNumber
        - subject
        - createdAt
      properties:
        ticketNumber:
          type: string
        subject:
          type: string
        createdAt:
          type: string
          format: date-time
    LeadInput:
      type: object
      required:
        - name
        - company
        - email
        - whatsapp
      properties:
        name:
          type: string
          minLength: 1
        company:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        whatsapp:
          type: string
          minLength: 1
        employees:
          type: ["integer", "null"]
        tier:
          type: ["string", "null"]
        tierPrice:
          type: ["number", "null"]
        subscription:
          type: ["number", "null"]
        totalReturn:
          type: ["number", "null"]
        language:
          type: string
          enum: [en, ar]
          default: ar
          description: Website language the visitor was using; the confirmation email is sent in this language
    Lead:
      type: object
      required:
        - id
        - name
        - company
        - email
        - whatsapp
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        employees:
          type: ["integer", "null"]
        tier:
          type: ["string", "null"]
        tierPrice:
          type: ["number", "null"]
        subscription:
          type: ["number", "null"]
        totalReturn:
          type: ["number", "null"]
        createdAt:
          type: string