# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Ticket {
  ticketNumber: string;
  subject: string;
  createdAt: Date;
}