# lib/api-zod/src/generated/types/ticketInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TicketInputLanguage } from './ticketInputLanguage';
import type { TicketInputTranscriptItem } from './ticketInputTranscriptItem';

export interface TicketInput {
  /**
     * @minLength 3
     * @maxLength 320
     */
  email: string;
  /**
     * @maxLength 200
     * @nullable
     */
  name?: string | null;
  /** @nullable */
  conversationId?: string | null;
  language?: TicketInputLanguage;
  /**
     * @minItems 1
     * @maxItems 60
     */
  transcript: TicketInputTranscriptItem[];
}